+++ packages/sources/dxfeed-test/src/endpoint/index.ts
@@ -1,8 +1,0 @@
-export { buildDxFeedHttpTransport } from './price'
-export {
-  EndpointTypes,
-  customInputValidation,
-  inputParameters,
-  endpoint as price,
-} from './price-router'
-export { buildDxFeedWsTransport } from './price-ws'